Python/week4_backend_ai/day22_fastapi/requirements.txt
# Day 22 - FastAPI Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic[email]>=2.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
